pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lakebench-k8s"
version = "1.0.0"
description = "Deploy and benchmark lakehouse stacks on Kubernetes"
readme = {file = "README.md", content-type = "text/markdown"}
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Andrew Sillifant" }
]
keywords = ["lakehouse", "spark", "iceberg", "kubernetes", "benchmark"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "boto3>=1.34.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0",
    "kubernetes>=29.0.0",
    "httpx>=0.27.0",
    "prometheus_client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.15.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "types-boto3>=1.34.0",
    "moto>=5.0.0",  # AWS mocking for tests
    "pre-commit>=3.0.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
lakebench = "lakebench.cli:app"

[project.urls]
Homepage = "https://github.com/PureStorage-OpenConnect/lakebench-k8s"
Documentation = "https://github.com/PureStorage-OpenConnect/lakebench-k8s/tree/main/docs"
Repository = "https://github.com/PureStorage-OpenConnect/lakebench-k8s"

[tool.hatch.build.targets.wheel]
packages = ["src/lakebench"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".gitignore",
    ".github/",
    "datagen/",
    "tests/",
    "conftest.py",
    "test-config*.yaml",
    "lakebench.spec",
    "lbrun.py",
    "Makefile",
    "install.sh",
    "docs/internal/",
    "docs/lakebench050sped.md",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["lakebench"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "kubernetes.*",
    "pyspark.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["lakebench.spark.scripts.*"]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (no external dependencies, fast)",
    "integration: Integration tests (require K8s/S3 connectivity)",
    "e2e: End-to-end tests (full deploy/run/destroy workflow)",
    "slow: Slow tests (skip with -m 'not slow')",
    "extended: Extended scale tests (scales 1, 10, 50, 100)",
    "stress: Stress tests at large scales (250, 500, 1000)",
]
